+++ kotlin-react/src/main/kotlin/react/Suspense.kt
@@ -1,0 +1,8 @@
+@file:JsModule("react")
+@file:JsNonModule
+
+package react
+
+external interface SuspenseProps : RProps
+
+external val Suspense: RClass<SuspenseProps>
